requirements.txt
fastapi~=0.116.1
uvicorn[standard]==0.30.0

# validation
pydantic==2.7.1
email-validator==2.3.0

#securitie
passlib[bcrypt]==1.7.4

#env
python-dotenv==1.0.1
#postgres
psycopg2-binary==2.9.9

# db
SQLAlchemy==2.0.30
sqlmodel==0.0.24

